GUVI HCL/requirements.txt
fastapi
uvicorn
pydantic
google-genai
requests
python-dotenv
flask
flask-cors
